// src/models/common.ts
export interface PaginationParams {
    _limit: number;
    _page: number;
    _totalRows: number;
}

export interface ListResponse<T>{
    data: T[];
    pagination: PaginationParams;
}

export interface listParams {
    _page?: number;
    _limit?: number;
    _sort?: string;
    _order?: 'asc' | 'desc';

    [key: string]: any;
}